requirements.txt
Django>=2.2
celery==4.4.1
redis==3.4.1
psycopg2>=2.7.5,<2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-celery-beat
django-celery-results
requests